Update getting_started.md
Added a few commas and full stops.

Section "Write the hello-world.js service"
Moved the first sentence of step 1 under the section heading as introduction for the task.
Reworded the second sentence of step 1 from "Choose the File -> New -> Project menu." to "Choose File -> New -> Project."
Reworded step 2 from "Enter the name of the project and click Create Project." to "Enter a name and create the project." to avoid describing details from the UI that are already obvious to the user.
Reworded step 3 from "Use right-click on the project icon in the Workspace Explorer to open the popup menu. Select `New -> Javascript Service`." to "Right-click on the project icon in the Workspace Explorer and choose `New -> Javascript Service`."
Step 4 combined two actions. So, I reworded step 4 from "Once you enter the name of the service you can open it in the editor with double click on the file icon." to "Enter a name for the service." for simplicity and clarity.
Step 5 did not include any action. So, I reworded step 5 from "It already contains the hello world service implementation." to "Double click on the file icon to open the file in the editor. The file already contains the hello world service implementation." by including a part of step 4 in it.
Reworded step 6 from "Now just publish the project by clicking on the `Publish` menu of the project." to simplify and emphasize on the action to be performed.
Reworded step 7 from "You can check in the Preview the result of the execution of our fancy server-side Javascript service, just by selecting the file icon in the Workspace Explorer." to "With the file icon selected in the Workspace Explorer, check the result of the execution of our fancy server-side Javascript service in the `Preview` tab." to unify the approach of each step starting with a call for action by the user.

Section "Update the hello-world.js" service"

Reworded step 2 from "The click Ctrl+S (Command+S for Mac) to save the file content." to "Save the file (Ctrl+S for Windows, Command+S for Mac)."
removed step 3 and included the text to the paragraph below.