Update getting_started.md
Changed "tab" with "view" and replaced "Workplace Explorer" with "Workplace" to unify terminology between this page and the ide_perspective_workbench.html page.

Replaced the javascript sample in step 5 with a screenshot showing the editor.

Changed expression "apart of" to "apart from" in the third bullet of the References section.

Changed word order in the sentence "The output changes immediately in the Preview view" to "The output in the Preview view changes immediately."

Removed one "more" from the sentence "If you would like to know more about how to write more complex services..." to prevent repetition.

Changed the expression "So far we saw how easy is..." to "So far we saw how easy it is..."

Reworded the sentence "More samples can be found here" to "You can find more samples here" to avoid passive voice.