Update architecture.md
Replaced "well proved" with "well-proved"
Added a few commas
Replaced the passive sentence "After the creation of the cloud application, it gets provided by the runtime components." with the more active "The runtime components provide the cloud application after you create it."
Replaced the definite article before "Java Web Profile compliant application server" with an indefinite one.
Added hyphenation to the compound adjective "Java-Web-Profile-compliant".
Removed ", there" from the phrase "On top, there are the..."
Divided the sentence "On top are the Eclipse Dirigible containers for service execution. , which depending on the scripting language and purpose can be: Rhino, Nashorn, V8, Mylyn, Flowable, ActiveMQ, etc." in two sentences and included a bulleted list.
Replaced "Rhino, Nashorn, V8" with "GraalVM JS".
Rephrased the sentence "Depending on the target cloud platform, in Eclipse Dirigible can be integrated the services provided by the underlying platform." to "Depending on the target cloud platform, you can integrate the services provided by the underlying technology platform in Eclipse Dirigible."