Update ide.md
Introduced the full version of the term Web IDE - web-based integrated development environment. Also, made corrections on other occurences of the term so it is consistent.
Rephrased "Three areas are with fixed positions" to "Three of the UI elements retain their positions in all perpectives".
Marked the three areas in the screenshot.
Replaced "minimal" with "smallest" in the phrase "minimal atomic parts".
Rephrased the para "The tools that constitute the perspectives are laid out in predefined regions of the work plot, but their position can be changed using drag and drop. The perspectives are simply predefined configurations, hence different tools may be opened, moved, or closed on the work plot of a perspective for the users’ convenience. The tools can also be maximized, minimized, or even popped out in a separate window." to address the user directly.
Added a few commas.
Replaced "Toolbar" with the already mentioned "top-area toolbar".
Formatted some portions of the text as UI text or code, depending on the case, to make it consistent with other pages.