Update concepts.md
Changed "must-to-know" with "must-know".
Reworded the sentence "There are several must-to-know concepts implied in the cloud toolkit, which have to be understood before getting started" to make it clear that it is the concepts that have to be understood, and not the cloud toolkit.
Changed "Repository" to "repository".
Replaced "Dirigible" with "Eclipse Dirigible".
Changed "Workspace" to "workspace".
Removed italics from a couple of words in the body text for consistency.
Changed "one only" to "only one".
Changed "API Management" to "API management trends" to make the sentence complete.
Changed "the strengths" to "their strengths".
Changed "the way of how services should behave as well as how to be described" to "the way services should behave and be described".
Removed "bit".
Changed "Entity Services" to "entity services" in the body text.
Reworded "There are several editors currently available that provide visual definition of application artifacts such as Database Schema Models, Entity Data Models and BPMN Processes." to "This is the visual definition of database schema models, entity data models, and BPMN processes. In Eclipse Dirigible, modelling is enabled by several [editors and modelers](editorsandmodelers.html)."
Added links to page editorsandmodelers.md
Added a few commas.