Update concepts_dynamic_applications.md
Removed caps from the expression "Dynamic Applications".
Removed caps from the expression "In-System Development".
Removed caps from the expression "Content-Centric".
Replaced "on" with "in" in the expression "stored on a generic repository".
Removed caps from the expression "Scripting Languages".
Moved "possible" at the end of the sentence.
Replaced "one" with "some".
Reworded "In general, a dynamic application consists of components, which can be separated into the following categories:" to "In general, components of a dynamic application consists can be separated into the following categories:"
Changed "well accepted" with "well-accepted".
Removed caps from the expressions "Data Structures", "Entity Services", "Scripting Services", "User Interface", and "Integration Services".
Added a few commas.
Updated the link to the API page in the "Scripting services" bullet.
Replaced "which" with "that" in the sentence "you can use a rich set of APIs, which give you access to the database and HTTP layer" since restrictive clauses are normally not set off with commas and included with "that".
Replaced "on" with "to" in the expression "bring on another level".
Replaced "which" with "that" in the sentence There are already many cool client-side AJAX frameworks, which you can use depending on the nature of your application."
Replaced "self-contained" with "self contained".
Changed "to create" to "for creating".