Update concepts_generation.md
Replaced "which" with "that" because of the restrictive clause.
Removed caps from "Entity Service".
Changed "is" to "are" in "Another example is user interface templates..."
Reworded "There can be provided also templates based on different frameworks for client-side interaction." to "Templates can also be provided based on different frameworks for client-side interaction."
Added an indefinite article "a" in the note.
Added a few commas and fullstops.
Changed "of" with "to" in the expression "In contrast of the approach above".
Removed caps from "Entity Data Modeler".
Replaced "for the generation of" with "to generate".