Update concepts_entity_service.md
Removed caps from "entity service" and made it italic.
Made "domain object management" italic.
Changed "used interface" to "user interface".
Reworded "The template uses as input a database table and an entity service name, which are entered in the corresponding wizard" to "As an input, the template uses a database table and an entity service name that are entered in the corresponding [wizard](../samples/entity_service.html)."
Reworded "From the context menu *Generate*, select *User Interface for Entity Service*." to "Choose `Generate -> User Interface for Entity Service`."
Applied different format to "[[protocol]://[host]:[port]/[dirigible application context]/services/js/[project]/[entity service path]" so its displayed in the web page.
Reworded "workspace explorer" to "workspace view" and reformatted.
Added a comma or a fullstop here and there.
Changed "which" to "that" in "a primary key, which will be used for its identity." because of the restrictive clause.
Changed "which" with "that" in "column types, which are supported by default" because of the restrictive clause.