Update concepts_rest.md
Reformatted "http-rs" to italics.
Removed caps from "Services".
Reformatted "hello-api.js" to italics.
Reformatted "Sending GET /services/js/v3/test/hello-api.js to the server, hosting the piece of code above in test/hello-api.js will return response body “Hello there!”." to "Sending a *GET* request

> GET /services/js/v3/test/hello-api.js

to the server and hosting the *hello-api.js* piece of code above in *test/hello-api.js* will return response body

> `Hello there!`  "

Changed "sample" with "example" to unify terminology.
Reformatted the example for root path ("").
Reformatted "Technically, configuration is not required to execute a service, but obviously it will do nothing, if you don't instruct it what to do. " to body text.
Divided "Now, this is a fairly simplistic example aiming to give you a hint of how you can bring up a REST API to life with http-rs and there is a whole lot more that we shall explore in the next sections." in two shorter sentences.
Added a few commas, fullstops, colons, etc.
Reformatted some of the method/service examples to unify with other pages.
Moved "perform a couple of more advanced activities, which will be reviewed in the [Advanced](#advanced) section below " to a bullet.
Replaced "such as" with "like" to avoid repetition.
Changed "Read below for more examples how to use." to "Read below for more examples how to use the methods."