Update concepts_web_content.md
Replaced the definite article with an indefinite in the first occurence of the expression "Web content adapter".
Changed "which" to "that" because of the restrictive clause.
Reworded "The default behavior of the adapter on a request to a whole collection and not to a particular resource is to send back an error code, indicating that folder listing is forbidden." to make it shorter and more clear.
Added a few definite and indefinite articles.