Update concepts_rest.md
Changed "rest" to "REST".
Changed the expression "fluent API" to "REST (rs-fluent) API" to introduce "rs" as an abbreviation used to refer to "REST".
Changed "rs fluent" to "rs-fluent" and removed "or" at the end of the bullet.
Unified format and language of the three bullets.
Divided second bullet into two separate sentences to improve clarity. Changed "fluent API" to "rs-fluent API" to improve antecedent basis. Changed "they are essentially the same thing" to "They are holding the configuration". Removed "and finally" at the end of the bullet.
Replaced "the third option is" with "You can". Changed "fluent API" to "rs-fluent API"
Changed "A configuration object is a JS object with canonical structure that rs can interpret" to "A configuration object is a JS object with canonical structure that the http-rs module can interpret"
Added a few commas and definite/indeifnite articles.